requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.0.0
python-multipart>=0.0.6

# Templates
jinja2>=3.1.2

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTTP client
httpx>=0.25.0

# RSS parsing
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# Scheduling
apscheduler>=3.10.0

# Settings & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Production
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

